test(convert): move docx callback tests to module
