+++ services/convert_service/src/process/mod.rs
@@ -1,2 +1,5 @@
 mod convert;
 pub mod runner;
+
+#[cfg(test)]
+mod test;
